remove impl and fix dump failed issue
